django_projects/institute/requirements/base.txt
# requirements/base.txt

mysqlclient==1.3.6